German: Sie sind Teil eines Teams, das eine Software fu ̈r die Verwaltung von Daten eines Kommunikationsnetzwerks erstellen soll. Neben dem Sammeln und Aufbereiten von Daten aus externen Quellen muessen in diesem Programm Abfragen zur Netzwerkinfrastruktur implementiert werden. Dazu muessen die Daten in einer geeigneten Datenstruktur abgelegt werden und darauf die Abfragen effizient ermoeglicht werden.
English:
You are part of a team, which is to create a software fu r managing
data of a communication network. In addition to collecting and
processing of data from external sources must be implemented for
network infrastructure in this program visits. These have to the data
in an appropriate data structure to be stored and then the queries are
made possible efficiently.